Turn weights into floats

+++ chem/models/base.py
@@ -74,7 +74,7 @@ class JCIBaseNet(pl.LightningModule):
         if weighted:
             weights = model_kwargs.get("weights")
             if weights is None:
-                weights = torch.sum(torch.cat([data.y for data in train_data]), dim=0)
+                weights = torch.sum(torch.cat([data.y for data in train_data]), dim=0).float()
                 weights = torch.sqrt(weights)
                 weights = 1 + torch.max(weights) - weights
                 mw = torch.mean(weights.float())
